requirements.txt
pandas>=2.3.0
numpy>=1.23.2
pyyaml>=6.0
python-dateutil>=2.8.2
pytz>=2020.1
tzdata>=2022.7 